deps: v2.29 step 12 DEP-TS - tree-sitter 0.27, TOKENIZER_REV 3, parser-only invalidation keeps trend (16-family cross-check byte-for-byte identical)
tree-sitter 0.26.11 -> 0.27.0 (both workspace lock files synced,
tree-sitter-language 0.1.8, NOTICE regenerated). `Node::child_count`
returns to u32: three conversions deleted in struct_fp.rs / units.rs /
tests submodule hs_grammar_pin.rs, and the scan/ast.rs header comment
states the widths of both counts (checked against the 0.27.0 source:
child/named_child take u32, named_child_count is still usize).

TOKENIZER_REV 2 -> 3. Cache invalidation splits into two layers - a
storage / algorithm key mismatch still rebuilds the whole database; only
a parser revision mismatch takes dedup/schema/parser.rs::invalidate:
clear the 13 parse-derived tables and the three stamps full_build /
resolve_key / mention_rev, roll the epoch, keep trend. The trend
toolchain stamp now carries tokenizer{REV}, and rows with an old stamp
are re-measured before reuse (schema_current = storage key AND parser
key; index::peek is still read-only; the meta key name is assembled in
exactly one place, parser::KEY). Decision: do not let one parser upgrade
erase the score history - trend rows are a cache rebuildable from git,
but rebuilding means measuring commit by commit (self repository 30
points, on the order of minutes), and carrying the parser revision in
the stamp is enough to separate "points from the old instrument" from
"points from the new instrument".

Invariance protocol (same tree 013f9fe, old / new binaries each with
their own worktree and .ce): the 16 families of JSON reports are
byte-for-byte identical - the churn difference on codex's first run =
the 14-day window moved forward between the two runs; re-run back to
back over the same window they are byte-for-byte identical (54361
bytes); the self repository's 12 index tables are row-for-row identical
(files 758 / fingerprints 38774 / symbols 9847 / sites 5082 / edges 3696
/ unitsig 9425 / docsegs 2074 / bag 190398 / df 5834 / mentions 336356 /
mention_files 1019 / result_cache 1); the mention reports for the four
external corpora are identical; the full-history FPR replay is identical
at 257 rows; the similar five-corpus replay is ok (the SQL reader and
the in-memory corpus agree bit for bit; tally is a function of the index
tables and SIMILAR_REV 1); ce check 945 / the ratchet drew nothing. Cold
ce dedup median 10545 -> 10232 ms, warm 668 -> 672 ms (three rounds),
and the database size is byte-for-byte identical. The shared index at
the main root ran a real parser-only invalidation once before this
commit: 761 files fully refreshed, no whole-database rebuild.

eval: tokens.rs lost its frozen anchor because a constant changed (the
three self-repository slices 25 -> 24); per the EVAL-SET.md revival
protocol, CE_REFREEZE=cli/src/dedup/tokens.rs continued the freeze by
name (graph-slice / t3-universe only swap the sha; for docdup-segments
that file gains one more live comment segment: the doc comment passed
the length floor, live 259 -> 260); the 25-row floor is restored.

Docs: booklet 01's invalidation sentence changed to "clear the
parse-derived tables, keep trend and re-measure it" and now cites
schema/parser.rs and the trend header comment (the tokens.rs:21 anchor
re-signed along with the constant's value); the PERF-BUDGET
invalidation-wording sentence; site how pages in both languages
TOKENIZER_REV 3; the README tree-sitter chip in both languages and the
facts projection blessed along with it; booklet 13's self-repository
census row U 1019->1022, unmentioned 9->6.

ADR-006 named re-establishment: main CHANGELOG 587->596,
dedup/schema.rs 274->284 (functions 11->12), dedup/tokens.rs 141->142,
new file into the baseline dedup/schema/parser.rs 55; tests submodule
it/main.rs 166->167, new files it/trend_parser.rs 33 /
unit/dedup/parser.rs 110 (it/docs_diagrams.rs 169->151).
Gates: main 945 / 55 / 0, tests submodule 984 / 119 / 0 (six legs green
in both roots), lib 373, full it 389 (12 ign), clippy + fmt clean (cli
--all-targets / gui --locked).

codex gpt-6-astra landed the code up to the point where the
cross-checks and the tests were all in place, then its subscription
quota ran out (restores Sep 13); Claude reviewed and closed out.
dependabot PR 5 is closed with this batch (the version bump landed in
the local lock files). The verdict, the score algorithm, the schema id
and the wire are all unchanged.

+++ cli/src/dedup/struct_fp.rs
@@ -47,7 +47,7 @@ pub fn spine(root: tree_sitter::Node) -> Spine {
         }
         // push children reversed so the stack pops them in order
         for i in (0..node.child_count()).rev() {
-            if let Some(child) = node.child(i as u32) {
+            if let Some(child) = node.child(i) {
                 stack.push(child);
             }
         }
