+++ cli/src/fourclass/units.rs
@@ -114,7 +114,7 @@ fn extra_units<'t>(
             out.push((unit, node));
         }
         for i in 0..node.child_count() {
-            if let Some(child) = node.child(i as u32) {
+            if let Some(child) = node.child(i) {
                 stack.push(child);
             }
         }
